test(query builder): split query builder tests
