+++ __test__/query-select.spec.ts
@@ -1,27 +1,17 @@
 import {
-  buildIndexExpr,
-  buildSelectExpr,
-  buildWhereClauseExpr,
+  connect,
+  IConditionExpr,
   ILetExpr,
-  IndexParamsOnExceptions,
-  MultipleQueryTypesException,
   InWithinOperatorExceptions,
+  ISelectType,
+  LogicalWhereExpr,
   Query,
-  QueryOperatorNotFoundException,
-  SelectClauseException,
+  QueryGroupByParamsException,
+  selectBuilder,
   SortType,
   WhereClauseException,
-  IndexParamsUsingGSIExceptions,
-  QueryGroupByParamsException,
-  ISelectType,
-  selectBuilder,
-  LogicalWhereExpr,
-  IIndexOnParams,
-  IIndexWithParams,
-  IConditionExpr,
-  connect,
 } from '../lib';
-import { connectionString, username, bucketName, password } from './testData';
+import { bucketName, connectionString, password, username } from './testData';
 
 const testQuery = async (query: string) => {
   const conn = connect({
@@ -33,38 +23,7 @@ const testQuery = async (query: string) => {
   return await conn.query(query);
 };
 
-describe('Test Query Types', () => {
-  test('Check select clause parameter types', async () => {
-    const dist: ISelectType = {
-      $distinct: {
-        $raw: {
-          $count: {
-            $field: {
-              name: 'ottoman',
-            },
-            as: 'odm',
-          },
-        },
-      },
-    };
-
-    expect(buildSelectExpr('', dist)).toStrictEqual('DISTINCT RAW COUNT(ottoman) AS odm');
-  });
-  test('Verify the exception throwing, if there is an error in the SELECT expression.', async () => {
-    const dist: ISelectType = {
-      $raw1: {
-        $count: {
-          $field: {
-            name: 'ottoman',
-          },
-          as: 'odm',
-        },
-      },
-    };
-    const run = () => buildSelectExpr('', dist);
-    expect(run).toThrow(SelectClauseException);
-  });
-
+describe('Test Query Builder SELECT clause', () => {
   test('Check query builder.', async () => {
     const select: ISelectType[] = [
       {
@@ -221,56 +180,6 @@ describe('Test Query Types', () => {
     expect(run).toThrow('The WITH clause has an incorrect syntax');
   });
 
-  test('Check the exception WHERE with an operator not found', async () => {
-    const expr_where = {
-      $nill: [{ address: { $like: '%57-59%' } }, { free_breakfast: true }, { free_lunch: [1] }],
-    };
-
-    const run = () => buildWhereClauseExpr('', expr_where);
-    expect(run).toThrow(QueryOperatorNotFoundException);
-  });
-
-  test('Check the exception WHERE', async () => {
-    const expr_where = {
-      $not: { address: { $like: '%57-59%' }, free_breakfast: true, free_lunch: [1] },
-    };
-
-    const run = () => buildWhereClauseExpr('', expr_where);
-    expect(run).toThrow(WhereClauseException);
-  });
-
-  test('Check WHERE clause parameter types', async () => {
-    const where: LogicalWhereExpr = {
-      $or: [{ price: { $gt: 1.99, $isNotNull: true } }, { auto: { $gt: 10 } }, { amount: 10 }],
-      $and: [
-        { price2: { $gt: 1.99, $isNotNull: true } },
-        { $or: [{ price3: { $gt: 1.99, $isNotNull: true } }, { id: '20' }] },
-      ],
-    };
-    expect(buildWhereClauseExpr('', where)).toStrictEqual(
-      '((price>1.99 AND price IS NOT NULL) OR auto>10 OR amount=10) AND ((price2>1.99 AND price2 IS NOT NULL) AND ((price3>1.99 AND price3 IS NOT NULL) OR id="20"))',
-    );
-  });
-
-  test('Check WHERE NOT operator', async () => {
-    const where: LogicalWhereExpr = {
-      $and: [
-        {
-          $not: [
-            { price: { $gt: 1.99 } },
-            { auto: { $gt: 10 } },
-            { amount: 10 },
-            { $or: [{ type: 'hotel' }, { type: 'landmark' }, { $not: [{ price: 10 }] }] },
-          ],
-        },
-        { id: 8000 },
-      ],
-    };
-    expect(buildWhereClauseExpr('', where)).toStrictEqual(
-      '(NOT (price>1.99 AND auto>10 AND amount=10 AND (type="hotel" OR type="landmark" OR NOT (price=10))) AND id=8000)',
-    );
-  });
-
   test('Check the WHERE clause function of the query builder', async () => {
     const expr_where: LogicalWhereExpr = {
       $or: [{ address: { $like: '%57-59%' } }, { free_breakfast: true }],
@@ -316,80 +225,6 @@ describe('Test Query Types', () => {
     );
     const execute = await testQuery(query);
     expect(execute.rows).toBeDefined();
-  });
-
-  test('Check the parameters of the INDEX clause', async () => {
-    const expr_where: LogicalWhereExpr = { 'travel-sample.callsign': { $like: '%57-59%' } };
-
-    const on: IIndexOnParams[] = [{ name: 'travel-sample.callsing', sort: 'ASC' }];
-
-    const withExpr: IIndexWithParams = {
-      nodes: [],
-      defer_build: true,
-      num_replica: 0,
-    };
-
-    const index = buildIndexExpr('travel-sample', 'CREATE', 'travel_sample_id_test', on, expr_where, true, withExpr);
-
-    expect(index).toStrictEqual(
-      'CREATE INDEX `travel_sample_id_test` ON `travel-sample`(`travel-sample`.callsing["ASC"]) WHERE `travel-sample`.callsign LIKE "%57-59%" USING GSI WITH {"nodes": [],"defer_build": true,"num_replica": 0}',
-    );
-  });
-
-  test('Check the INDEX clause of the query builder', async () => {
-    const expr_where: LogicalWhereExpr = { 'travel-sample.callsign': { $like: '%57-59%' } };
-
-    const on: IIndexOnParams[] = [{ name: 'travel-sample.callsing' }];
-
-    const withExpr: IIndexWithParams = {
-      nodes: ['192.168.1.1:8078', '192.168.1.1:8079'],
-      defer_build: true,
-      num_replica: 2,
-    };
-
-    const query = new Query({}, 'travel-sample')
-      .index('CREATE', 'travel_sample_id_test')
-      .on(on)
-      .where(expr_where)
-      .usingGSI()
-      .with(withExpr)
-      .build();
-
-    expect(query).toStrictEqual(
-      'CREATE INDEX `travel_sample_id_test` ON `travel-sample`(`travel-sample`.callsing) WHERE `travel-sample`.callsign LIKE "%57-59%" USING GSI WITH {"nodes": ["192.168.1.1:8078","192.168.1.1:8079"],"defer_build": true,"num_replica": 2}',
-    );
-  });
-
-  test('Check Multiple Query Exceptions with select', () => {
-    const run = () => new Query({}, 'travel-sample').index('travel_index', 'CREATE').select('*');
-    expect(run).toThrow(MultipleQueryTypesException);
-  });
-
-  test('Check the DROP INDEX clause of the query builder', async () => {
-    const query = new Query({}, 'travel-sample').index('DROP', 'travel_sample_id_test').usingGSI().build();
-
-    expect(query).toStrictEqual('DROP INDEX `travel-sample`.`travel_sample_id_test` USING GSI');
-  });
-
-  test('Check the exception Index Params On Exception', async () => {
-    const run = () => new Query({}, 'travel-sample').select().on([{ name: 'test', sort: 'DESC' }]);
-    expect(run).toThrow(IndexParamsOnExceptions);
-  });
-
-  test('Check the exception Invalid index name', async () => {
-    const run = () => new Query({}, 'travel-sample').index('CREATE', 'index-*&');
-    expect(run).toThrow(
-      'Valid GSI index names can contain any of the following characters: A-Z a-z 0-9 # _, and must start with a letter, [A-Z a-z]',
-    );
-  });
-  test('Check Multiple Query Exceptions with index', () => {
-    const run = () => new Query({}, 'travel-sample').select('*').index('travel_index', 'CREATE');
-    expect(run).toThrow(MultipleQueryTypesException);
-  });
-
-  test('Check the exception IndexParamsUsingGSIExceptions', () => {
-    const run = () => new Query({}, 'travel-sample').select('*').usingGSI();
-    expect(run).toThrow(IndexParamsUsingGSIExceptions);
   });
 
   test('Check the exception WITH clause', () => {
